Create subdirectory for portable source code.
The idea is to split the platform-specific code off into subdirectories, while
keeping the common code together. Currently, there are four subdirectories:

  - blake3: the BLAKE-3 source code
  - leopard: the Leopard O(n log n) FFT-based Reed Solomon implementation
  - libpar3: the par3 library code (uses blake3 and leopard)
  - par3cmd: the par3 command line tool (uses libpar3)

This code was trivially moved from the existing linux/ subdirectory which
was a copy of the windows/ subdirectory.

I plan to introduce separate subdirectories for Windows and Linux (POSIX)
specific code later.

It's currently possible to build a Windows executable on Linux using
the CMake configuration and the MinGW compiler toolchain (see
portable/BUILDING.txt for details).