Port par3cmdline to Linux.
This is based on the current Windows implementation, but uses a different
philosophy than the existing (incomplete) Linux port: instead of littering
the par3 source code with #if __linux__ / #elif _WIN32 conditionals to
define alternate versions of platform-specific functions, all the
platform-specific code is moved to the platform/ subdirectory.

So we have the following subdirectories:

  * portable/blake3           -- third-party Blake3 library
  * portable/leopard          -- third-party Leopard-RS library
  * portable/libpar3          -- portable PAR3 code
  * portable/par3cmd          -- par CLI tool (main.c)
  * portable/platform         -- definitions of platform-specific functions
  * portable/platform/linux   -- linux implementations
  * portable/platform/windows -- windows implementations

I used CMake as the meta-build system, since it seemed easier to get working
than either autoconf or meson. See portable/BUILDING.txt for instructions on
how to build.

The hope is that this easier to maintain, and will make it easier to port
par3 to additional platforms.

Since the par3 code is heavily based on Windows APIs, the Linux port is
mostly emulating Windows CRT functions at this point. There is room for future
improvement here, by defining more abstract APIs that are a better fit for
all platforms (e.g., the findfirst()/findnext() API uses Windows file
attributes that do not map neatly to POSIX file modes.)

I moved the source files from the linux/ subdirectory (which were identical
to the source files in windows/) to keep the git diff minimal and because my
ultimate goal is to only have 1 copy of each file in the repository. I kept the
remaining files in the linux/ directory for now, because they contain some
useful documentation on how to (cross)compile with autoconf.

I kept the windows/ directory entirely intact because it contains a Visual
Studio project. Ideally that should be recreated in the portable/ subdirectory
instead, and then the windows/ subdirectory can be removed, so that each source
files exist only in one place.

There is still a lot of additional work that could be done to bring the code
up to production quality:

  - Support for multithreading (OpenMP for Leopard, ??? for Blake3)
  - Support for installing the binary (make install)
  - Support for installing as a library with a header (for third-party dev)
  - Automated testing: unit tests & integration tests would be nice
  - Additional compiler flags like -fhardened to improve security
  - Fix compilation warnings

etc.